#566 Refactored blog
- Moved types and DTOs from /api/blog-api.tsx to constants/types/blog-types.ts
- Replace dynamic positioning of the UserSummary through inline styling with tailwind
