+++ quolance-ui/src/components/ui/blog/CreatePostModal.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useRef, useEffect } from 'react';
 
 interface CreatePostModalProps {
     open: boolean;
