+++ quolance-ui/src/constants/types/blog-types.ts
@@ -19,3 +19,35 @@ export type CommentType = {
     content: string;
     dateCreated: string;
 }
+
+export interface CommentResponseDto {
+    commentId: string;
+    blogPostId: string;
+    username: string;
+    content: string;
+  }
+  
+export interface CommentRequestDto {
+content: string;
+}
+
+export interface BlogPostUpdateDto {
+    postId: string;
+    title: string;
+    content: string;
+    // tags: string[];
+    // files?: File[];
+  }
+
+  export interface ReactionResponseDto {
+    id: string;
+    reactionType: string;
+    userId: string;
+    userName: string;
+    blogPostId: string;
+  }
+  
+  export interface ReactionRequestDto {
+    reactionType: string;
+    blogPostId: string;
+  }
